Soft Limit for Climber
The first thing I did was I first created the initialization for the climber's integrated sensor. This is on line 32. Then I created a method that would check if the climber's position is greather than the soft limit constant that was set in the Robot Constants class, and it also took in the climber's current and checks if it is above four. If the current is above four and the position is above the soft limit, the climber's position is set to 0 so it does not continue moving.